[RFC] feat(dbt): move scaffold to separate files (#15723)
## Summary & Motivation
Resolves https://github.com/dagster-io/dagster/issues/15603.

Because building a schedule using `@dbt_assets` and
`build_schedule_from_dbt_selection` requires the user to have an
instance of the `AssetsDefinition`, it is not enough to only separate
out the schedules to `schedules.py`. More separation must occur,
otherwise the user will need to deal with circular dependencies.

We make these changes. Now the following scaffold structure is as
follows (e.g. for a scaffolded `jaffle_shop_dagster` project):

```bash
❯ tree jaffle_shop_dagster
jaffle_shop_dagster
├── jaffle_shop_dagster
│   ├── __init__.py
│   ├── assets.py
│   ├── constants.py
│   ├── definitions.py
│   └── schedules.py
├── pyproject.toml
└── setup.py

1 directory, 7 files
```

## How I Tested These Changes
bk