Allow setting config at the executor level that applies to all steps launched by the k8s_job_executor
Summary:
If users want to set some configuration to every step pod when using the k8s executor, their only recourse is the clunky solution of creating an op factory and using it on every one of the ops. A more natural place is to put this as config on the executor. Do that.

Test Plan: BK