[Run Tags] Pass original key to action (#15858)
## Summary & Motivation
Noticed that the op_selection tag wasn't able to be hidden/shown, this
is because it is renamed from solid_selection and the action was
receiving the renamed key instead of the original key. This updates the
logic to pass the original key instead.

## How I Tested These Changes
Locally toggled the op_selection tag.