[brownfield-essentials] socket support for ExternalExecutionTask (#15849)
## Summary & Motivation

Add support for TCP socket IO mode to `dagster-external`. This can be
set as either `input_mode`, `output_mode`, or both. In all cases, the
orchestration process spins up a TCP server. The same server can be used
for both input and output.

This also opens the door to allowing the external process to query the
orchestration process during the run, if we want to support that.

With the many possible IO modes `ExternalExecutionTask.run` has gotten
fairly complex. Currently it uses a "flat" conditional structure, but I
think it should be refactored to use context managers later.

## How I Tested These Changes

New unit tests