[brownfield-essentials] make sure output stream is closed when dagster external process shuts down (#15889)
## Summary & Motivation

This was hanging on 3.11. Turns out the problem was a [breaking
change](https://blog.pecar.me/python-enum) to how enums are
string-formatted in Python 3.11, which was breaking the socket protocol
(which uses an enum to define control messages). Also added an `atexit`
call to guarantee client socket gets closed cleanly, though this turned
out not to matter.

## How I Tested These Changes

Existing test suite with `test-all` (tests all python versions).