[ui] Asset feature context (#15878)
## Summary & Motivation

Introduce `AssetFeatureContext`, which allows injecting tabs and views
into the OSS assets view, so that we can add Cloud-specific features to
this part of the Dagster app.

- Refactor the Assets tabs code so that it is build separately from
`AssetView`. Configuration is provided via a tab builder function that
can be overridden.
- Similarly, refactor how we render the selected tab view to allow
defining a fallthrough view.

In Cloud, we can then define a provider that uses Cloud-specific tab and
view configuration.

## How I Tested These Changes

View an asset in OSS app, verify that current rendering and behavior
appears unchanged. Tabs and views all work correctly.

View an asset in Cloud app, verify that context overrides are consumed
correctly, with new tabs and view rendering as expected -- as long as
the relevant feature gate is enabled. See corresponding Cloud PR for
details.