feat(dbt): convert `Path` to `str` for `DbtCliResource` (#15901)
## Summary & Motivation
This was also noted in
https://github.com/dagster-io/dagster/issues/15689, but a long term fix
would require some sort of mapping from `Path` to `str` in our config
library.

As an example, a workaround that some users have done is:
https://github.com/dagster-io/dagster/issues/14564#issuecomment-1576830733,
which does just that.

Although this change is still bothersome for users who have Python type
checkers in place, at least this won't error if they choose to use
`Path` in place of `str`.

## How I Tested These Changes
pytest