[docs] Provide an asPath value for unversioned docs (#15912)
## Summary & Motivation

The unversioned `/migration` and `/changelog` do not currently have
breadcrumbs rendered server-side because I didn't specify their paths
for hierarchy lookup. This means that when the crawler tries to
determine the hierarchy for the page, it can't.

This fixes the issue.

## How I Tested These Changes

`yarn build`, then `yarn start`. Verify that when I request `/migration`
in the browser, the breadcrumbs are present in the HTML response.