[CRA -> Nextjs] Use Nextjs to build OSS app (#15842)
## Summary & Motivation
We're consolidating on next.js across all of our applications. This
brings OSS in line. Counterpart to the cloud pr:
https://github.com/dagster-io/internal/pull/6356

I had a hard time getting the types for svgs to use the Next type
definitions. For some reason typescript thinks that svgs are imported to
strings despite the `svg.d.ts` declaring them as `any`. I'm not sure
where that type is coming from but I did some `any` hacking to work
around that issue for now.

Changes to how we develop:
- Nothing changes for now. 
  - The next step will be to move OSS to Next.js
  - Then we can move all of our routes to use Next.js App Router

Chopping block:
- No more react-scripts, .dagster.js, or CRA 🥳 
- CSP Policy, we have to maintain it manually rather than generating it.
It lives in the root of the project as `csp-header.conf`

## How I Tested These Changes

1. `yarn build` followed by `dagster dev`
2. NEXT_PUBLIC_BACKEND_ORIGIN="http://127.0.0.1:3333" yarn start
3. yarn storybook still works in the `ui-components` and `ui-core`
packages