Wait for grpc calls to finish before shutting down the server on termination
Summary:
This is a prereq for supporting sensor ticks that are longer than 60 seconds b/c of the issue in https://github.com/dagster-io/dagster/issues/14523 - the daemon reloads servers locally every 60 seconds, and stops heartbeating on the old ones - so the grpc server shuts down, forcibly terminating any in-progress RPCs.

Luckily the grpc server APIs give you a way to shut down more gracefully and wait for rpc handlers to finish, we just weren't using it properly and were shutting everything down after 5 seconds: https://grpc.github.io/grpc/python/grpc.html#grpc.Server.stop

Increase that call to whatever the client timeout is to ensure that any calls that we have given up on on the server, the client has too.

Test Plan:
Write a sensor that takes 3 minutes to tick, set DAGSTER_GRPC_TIMEOUT_SECONDS to 300, run dagster dev with the sensor on
Before: After ~90 seconds the server would shut down and the tick would be interrupted with an UNAVAILABLE error
Now: The tick completes
Do the same but with DAGSTER_GRPC_SHUTDOWN_GRACE_PERIOD set to 5, see the same behavior as the "Before:" case above