[docs] - Essentials v1 sidenav updates (#15885)
## Summary & Motivation

This PR updates the sidenav in the docs to reflect the changes in the
Essentials v1 project:

- Adds a **Concepts > Advanced** section
- Moves some concepts to **Advanced**: Ops, Graphs, Jobs, I/O, Dagster
Types, Logging, and GraphQL
- Removes URLs from sidenav items that have children, but don't have a
category page. Ex: Partitions & backfills. It feels like a weird
experience to click on an item like this and be taken to a page that
only describes one of the items. This is simpler than adding adding
category pages, at least for now.

## How I Tested These Changes

👀 , local