Example Asset-based YAML DSL (#15744)
## Summary & Motivation

This is an example demonstrating how to build a YAML DSL for asset graph
construction. This is very much a WIP and will be refined and
documented. The idea here is to provide an example that users an copy
and paste and then customize to their own desires, as well as provide a
testing ground for dagster externals.

## How I Tested These Changes

Included unit tests and `dagster dev`

---------

Co-authored-by: Sean Lopp <lopp@elementl.com>