[externals] Make dagster-external no-op if launching process was not dagster (#15957)
## Summary & Motivation

This allows scripts using `dagster-external` to not crash if the script
was not launched by a dagster `ExternalExecutionTask`.

- Add a new injected env var `DAGSTER_EXTERNAL_IS_ORCHESTRATION_ACTIVE`
- Add a top level function `is_dagster_orchestration_active` that checks
this env var.
- If this env var is not set, and `init_dagster_external` is called,
emit a warning and return a `unittest.mock.MagicMock` object in place of
the context. This will cause any methods called on the context to just
return another `unitest.mock.MagicMock` object.

Obviously the user will still need to add special logic to handle the
"not-launched-by-dagster" case if using context info in any significant
way, but this allows `init_dagster_context` to be called without
crashing.

## How I Tested These Changes

New unit test.