feat(dbt): recommend using `dbt.log` when viewing errors (#15958)
## Summary & Motivation
When encountering errors when running dbt with Dagster, it takes time to
be intuitive to search through terminal stdout to parse the dbt logs or
see the logs in Dagster's compute logs in the Dagster UI.

In dbt, a `dbt.log` file is written out for each dbt command invocation.
Here, when an exception occurs, we now also recommend the user view the
contents of that file to debug. Like dbt artifacts, to ensure that the
logs are retained for each command invocation, ensure that
`DBT_LOG_PATH` is set so that each log file has a unique path.

See
https://docs.getdbt.com/reference/global-configs/logs#log-and-target-paths
for more details.

## How I Tested These Changes
pytest

Co-authored-by: Rex Ledesma <rex@elementl.com>