[brownfield-essentials] Refactor ExternalExecutionTask to better support multiple environments (#15892)
## Summary & Motivation

This is another iteration of the external execution API, stacking on and
motivated by @alangenfeld's `DockerExecutionTask` in #15820. The goal
here is to move toward base external execution APIs that can easily be
extended to different environments.

- Split `ExternalExecutionTask` into:
- `ExternalExecutionTaskBase`: an abstract base class that provides
extension points at `_input_context_manager`, `_output_context_manager`,
and `_launch`.
- `ExternalExecutionSubprocessTask`: implements
`ExternalExecutionTaskBase` for the subprocess case.
- Refactor `DockerExecutionTask`:
    - Inherit from `ExternalExecutionTaskBase`
- Instead of overriding `run`, override `_launch`,
`_input_context_manager`, `_output_context_manager`

### Detailed explanation of current design

Let's call the combination of the `dagster-external` library and
associated set of orchestration-side abstractions the "Dagster External
API". An "adapter" is an environment-specific implementation of the
Dagster externals API. Currently there are two adapters in master:
subprocess and Docker. In these PR, these are each defined with similar
formatting in dedicated modules
(`dagster._core.external_execution.subprocess` and
`dagster_docker.external_resource`).

An adapter should define the following parts:

- `AdapterExecutionTask`: A class inheriting from abstract base class
`ExternalExecutionTask` (e.g. `SubprocessExecutionTask`)
- `AdapterExecutionResource:` A resource inheriting from
`ExternalExecutionResource` (e.g. `SubprocessExecutionResource`)
- `AdapterTaskParams`: A dataclass inheriting from `ExternalTaskParams`
(e.g. `SubprocessTaskParams`)
- `AdapterTaskIOParams` An optional dataclass inheriting from
`ExternalTaskIOParams` (e.g. `SubprocessTaskIOParams`)

The user-facing component of an adapter is the
`AdapterExecutionResource`. This exposes a method `run` that is
responsible for taking a flat list of user-provided parameters,
instantiating an `AdapterExternalTask`, and calling `run` on this
instance. `run` returns once the `AdapterExternalTask` has completed
successfully and all associated resources have been cleaned up. `run`
throws an error if `AdapterExternalTask` is not successful.

In order to provide common infrastructure for all adapters, the
`AdapterExecutionTask` uses different channels for parameters
standardized across the Dagster Externals API ("API parameters") and
those that are specific to an adapter ("adapter parameters"). This
distinction is immaterial to the user, who mixes these parameters
together at the level of the configuration of `AdapterResource` and call
to `AdapterResource.run`.

The`AdapterResource.run` method makes the distinction, partitioning the
params into the API parameters and adapter parameters. The API
parameters are passed to the `AdapterExecutionTask` constructor. The
adapter parameters are packaged into an instance of `AdapterTaskParams`
and passed to `AdapterExecutionTask.run`.

The adapter task is responsible for launching the external task,
supplying it with orchestration context info, and reading notifications
streamed back during process execution. Most of the plumbing for this is
provided by the base `ExternalExecutionTask`. Adapter tasks plug in by
defining three required methods:

- `_input_context_manager`: returns a context manager that wraps
`_launch` and handles IPC input (i.e. writing context) to the external
task. This context manager must yield an instance of
`AdapterIOTaskParams`, which at a minimum contains environment variables
that `dagster-external` uses to establish communication with the
orchestration process.
- `_output_context_manager`: returns a context manager that wraps
`_launch` and handles IPC output from the external task. Just like the
input context manager, this context manager also yields an instance of
`AdapterIOTaskParams`.
- `_launch`: launches the external process and monitors it for
completion. It returns if execution was successful, throws an error if
not (e.g. a subprocess with non-zero exit). It does not need to handle
any I/O or associated resource cleanup-- this is done by the base-class
provided plumbing and adapter input/output context managers. `_launch`
receives the `AdapterTaskParams` instance provided by the calling
`AdapterExecutionResource` as well as the two `AdapterIOTaskParams`
instances yielded by the context managers. It may use arbitrary logic to
combine these parameters to configure the external system.

**NOTE**: @alangenfeld has suggested composition as an alternative to
inheritance, and I'm not opposed to that. I thought about trying to
implement it but the path forward wasn't clear, whereas moving to an ABC
was a straightforward improvement over the status quo. We can continue
iterating.

## How I Tested These Changes

Existing test suite.