[docs] Show breadcrumbs on every doc page (#15916)
## Summary & Motivation

The Algolia crawler is looking for `DocSearch-lvl0` elements to assign
the top level of hierarchy, but on root pages of sections (e.g.
`/tutorial`) we don't show the breadcrumbs at all, which means for these
pages we fall back to "Dagster Docs" as the level-0 hierarchy value.

To improve this, always show breadcrumbs, even if it's just one item.
This looks a bit redundant on the root pages, so I tried to make the
`h1` elements a little more descriptive of what's actually on those
pages. Feel free to push back on this or tell me better strings to use!

## How I Tested These Changes

View docs site, verify that breadcrumbs always appear, and that modified
strings appear as expected.