Reduce lru_cache size on get_partition_keys_in_time_window from 100 to 5
Summary:
As we have been digging into memory on auto-materialization policies a bit more I realized that I hadn't really been considering the memory size of the cached result here, particularly since the time window is updated on each iteration right? So if there are 70k partitions, we'll be keeping, say, up to 100 2MB chunks. Seems like a lot - since reducing the cache to 5 helps with the specific test case we have been optimizing around, let's be conservative and drop it down to 5 (and maybe it's still worth investigating that instance-scoped partitions cacher class to pass around).

Test Plan: BK, run a dry run reconciliation on a large set of assets with many hourly partitions