[ui] Address significant (20s+) wait when opening Materialize modal (#15994)
## Summary & Motivation

Before vs After loom!
https://www.loom.com/share/98e9ef0a95ff4f15b05a9a3feed08704

This PR speeds up the rendering of the "Choose Partitions" modal in a
few ways:

1) When the `usePartitionHealthData` hook is told it can refresh in the
background as partitions or partition data changes, it attempts to
populate on mount using data from the Apollo cache. This means that on
the Asset partitions page, which shows the health bar, clicking
Materialize doesn't re-fetch the partition health all over again from
scratch before you can do anything.

2) A new helper (`keyCountInSelections`) optimizes a few code paths that
need the COUNT of selected partitions only, and were previously building
a huge array to call .length.

3) The Choose Partition modal was using an extremely inefficient
strategy to filter for "failed and missing only", which iterated over
the selected partition keys and called through to code that ran an
`indexOf(key)` each time (even though the keys were checked
sequentially) and even worse, it was running the logic to do so even
when the checkbox was unchecked. I exposed a new getter that takes the
key indexes to remove the indexOf. This takes it from O(large^2) to
O(large) and is way faster.

Some of this code is still `O(dimension1 * dimension2 * range count)`
for 2D assets which is **_not great_**, but with these fixes applied it
is sufficiently fast for a 200k partition key set and a similar ~50 x
5000 partition key set. There are two primary reasons this code is so
messy. A) it supports multi-selection of assets (where partition health
ranges are unioned and there can be >1 status per partition), and B) our
system does not send down MISSING ranges. ("Missing" is the absence of a
range).

I think the latter could be changed fairly easily and would help to
clean up some of this a fair bit. "Failed ranges + anywhere there isn't
a range in this 2d space" is hard to get right and requires merging the
results. (See `AssetPartitions.tsx::dimensionKeysInSelection`)

## How I Tested These Changes

We thankfully have extensive test coverage of all this! I also took
screenshots of the UI before and after these changes to verify I see the
same things for all my test assets.

---------

Co-authored-by: bengotow <bgotow@elementl.com>