[automaterialize] Update GQL + UI to use GrapheneAutoMaterializeRule / GrapheneAutoMaterializeRuleEvaluation (#15996)
## Summary & Motivation

@bengotow notes:

This PR swaps out the existing GraphQL schema for auto-materialization
`conditions` for a new schema, which exposes the concept of
auto-materialization "rules" (attached to the policy) and rule
evaluations, which occur on each tick. This maps really naturally to the
existing UI. Instead of the UI mapping the `conditions` data into
specific hard-coded rows on the page, the entire center column can be
displayed using the data from the GraphQL API.

The idea behind this change is that we can more easily add new
auto-materialize rules + change rule descriptions without needing to
update the front-end. One example of this is the `Exceeded 5
materializations per minute limit` rule. The number 5 is now part of the
rule description sent from the backend and doesn't need to be blended in
from another fetch in the UI.

## How I Tested These Changes

I viewed both partition and un-partitioned assets with a variety of
ticks and took screenshots of the UI before + after. I also updated the
storybooks + mocks to reflect the new schema and verified that they look
good.


<img width="1840" alt="Screen Shot 2023-08-22 at 7 07 52 PM"
src="https://github.com/dagster-io/dagster/assets/1037212/0fbb4762-65a8-4067-b452-7a799364e990">

---------

Co-authored-by: bengotow <bgotow@elementl.com>