Surface asset check run failures (#16102)
Per the write path described in
https://github.com/dagster-io/dagster/pull/16006, the status goes

- planned event
- row created with status PLANNED
- evaluation event
- row updated with status SUCCESS/FAILURE

but there's also a case where

- planned event
- row created with status PLANNED
- run failure

and the check stays in PLANNED status. Similarly to how we reconcile
asset status at read time, use the status of the associated run to
adjust the status when we surface it over gql.

---

In the future with both assets and asset checks, I'd like to move to
emitting explicit MATERIALIZATION_FAILURE/ASSET_CHECK_RUN_FAILURE events
when the step fails (or when the run monitoring daemon fails the run).
Plus analogous skip or cancel events. This would mean
- we can keep actual status in the db. helpful for filtering on status
etc.
- we can use the event log consumer to alert on materialization and
check run failures

but that's a bit larger of a project, and I think it's safe to mimic
asset behavior in the meantime