[bugfix] Input asset record resolution for data versions of step-internal dependencies (#16058)
## Summary & Motivation

Resolves #15822

This is an attempt to fix a race condition described in:
https://elementl-workspace.slack.com/archives/C03A0D72A6T/p1692824586523789

When there are step-internal asset dependencies, the data version
tracking machinery needs to retrieve new materialization events created
_during_ a step in order to add accurate upstream dependency storage ids
to materialization records. Prior to this PR, this was done by querying
`instance.get_latest_data_version_record` in a loop, waiting for the
returned data version to match the cached data version from the current
run (indicating it is the correct storage ID).

This caused a race condition where multiple runs materializing the same
asset could cause the "latest" version to never match the cached
version, if a paralell run materialized the same asset.

This PR changes resolution logic to just always used the most recent
materialization, regardless of whether it was created in-step. It emits
a warning if it does not match the step-internal cached data version
This is acceptable because it is ambiguous which data version is
actually correct (depends on the external multi-asset step execution
machinery). We will need to figure out a more robust solution in the
future.

## How I Tested These Changes

Added new unit test.