[externals] Pluggable orchestration-side context-source/message-sink (#16169)
## Summary & Motivation

This is an RFC aiming to establish orchestration-side composability of
IO and launch.

- Establish a standard interface for context source and message sink on
the orchestration side
- Both `ExternalExecutionContextInjector` and
`ExternalExecutionMessageReader` are callables that take an
`ExternalExecutionOrchestrationContext` and produce a context manager
that yields a dict of params
- `ExternalExecutionResource.run` now includes `context_source` and
`message_sink` in signature (alongside `context` and `externals`
- Add `ExternalExecutionEnvContextSource` that sends context encoded
directly in environment variable
- Subprocess tests have been modified to use this alternative source.
- Add base64 encoding to JSON env vars (necessary for databricks and I
assume other platforms)
- Rename `ContextSource` to `ContextInjector` (orch side) and
`ContextLoader` (ext side)
- Rename `MessageSink` to `MessageReader` (orch side) and
`MessageWriter` (ext side)

The above changes aim at two goals:

- Allow context sources/message sinks to be freely composed with
orchestration-side resources (i.e. launchers).
- Decouple IO params from the environment. Context sources and message
sinks yield an arbitrary dict of params (containing e.g. `path` for the
file variants) to the launcher. It is up to the launcher to encode the
params and send them over the wire. All of the current launchers do this
by passing them as env vars, but the door is now open for launchers to
use a different channel.

This also gets into the "currying" issues Alex raised a few days ago.
Sources/sinks need to have some params supplied by the user (e.g. a
bucket/key for S3) and the context supplied by the framework. The
solution here is for the accepted interface for both to be a
`Callable[[ExternalExecutionOrchestrationContext],
ContextManager[Mapping[str, Any]]`, i.e. this:

```python
@contextmanager
def context_source(
    context: "ExternalExecutionOrchestrationContext",
) -> Iterator[Mapping[str, Any]]:
    ...
```

The builtin file-based ones defined the above in a closure that binds
`path`-- I am using this pattern to implement s3 source/sink upstack as
well.

## How I Tested These Changes

Modified existing tests