[auto-materialize] Add implicit AMPs to tests (#16249)
Previously, in asset reconciliation tests, we would infer the auto
materialize policy for assets that did not have one defined. This PR
adjusts behavior to instead assert that an auto-materialize policy is
defined on an asset that we expect to yield an evaluation for. Legacy
reconciliation tests with no AMPs now infer the AMP at scenario
construction time.

This also improves testing as AMPs now exist on the asset definition
(instead of being inferred) and ensures we aren't evaluating assets with
no AMP.

Necessary for: https://github.com/dagster-io/dagster/pull/16250