[externals] Interface for loading external-side params (#16222)
## Summary & Motivation

- Add `ExternalExecutionParamLoader`, an ext-side interface for
extracting params consumed by `ExternalExecutionContextLoader` and
`ExternalExecutionMessageWriter` from "the environment" (defined broadly
as the runtime context). A `param_loader` can be passed to
`init_dagster_externals` alongside a `context_loader` and
`message_writer`. The default is to read from environment variables.
- Remove `ExternalExecutionContextLoader.scoped_context`, make
`load_context` a context manager that takes `ExternalExecutionParams`.
- Factor out an additional `ExternalExecutionMessageWriterChannel` class
from `ExternalExecutionMessageWriter`. `ExternalExecutionMessageWriter`
now has an `open` context manager method that takes
`ExternalExecutionParams` and yields an
`ExternalExecutionMessageWriterChannel`.

This is an ext-side solution to the "currying problem" where we need to
run loader/writer initialization routines for two sets of parameters:

1. Params needed to instantiate the class on the ext side (e.g. a write
interval for blob message writers)
2. Params passed over from the orchestration side (e.g. the S3 bucket to
write to)

Prior to this PR, the means of reading type (2) was coupled to a
loader/writer class. This PR decouples the functionality such that
loader/readers need not concern themselves with _how_ parameters are
passed from the orch side-- that is handled separately by the
`ExternalExecutionParamLoader`. I assume in the majority of cases users
won't need to worry about this at all because the default
environment-variable-based implementation will be sufficient.

## How I Tested These Changes

Existing test suite.