Optimize the 'quick check' in get_asset_partitions_updated_after_cursor when there are a large number of partitions for a given key (#16213)
Summary:
For assets with tons of partitions, it is not uncommon to end up in a
situation where you have a single asset partition being passed in, that
is then compared to a dictionary with a ton of partition keys. The old
logic here was iterating through the giant dictionary and comparing each
element with the passed in partition key. Doing the reverse when
possible makes the function much faster in this case (the original logic
was only ~20-30ms per call, but the fact that we were then doing it once
for each partition in certain cases made it basically O(n^2) in the
number of partitions).

Test Plan: Working on an asset perf scenario
Manually, a single asset rule that was previously taking ~20 minutes now
takes ~30 seconds