Introduce limit for stale resolution for self-partitioned assets (#16285)
## Summary & Motivation

Self-partitioned assets can cause a recursion error during stale status
resolution if there are too many partitions. This is because resolution
of the Nth partition will recurse back to the 1st partition, since this
is within its ancestor tree.

This PR adds a threshold number of dependencies beyond which no
staleness status will attempt to be derived for self-partitioned assets.
This follows the existing approach of opting out of staleness
calculation for partitions with large numbers of "horizontal"
dependencies.

See:
https://elementl-workspace.slack.com/archives/C03D094F23G/p1693573479147479

## How I Tested These Changes

New unit test.