[ui] Add feature context for Job page (#16256)
## Summary & Motivation

Similar to the newly added "asset feature" context provider, this adds a
"job feature" context provider that allows setting the tabs on the Job
page, as well as setting a fallthrough route to handle custom pages on
the Job permalink.

## How I Tested These Changes

Verified that the Job page loads successfully in OSS. Same for snapshot
pages. Navigate to tabs, verify that routes load properly.

Verified that it also loads successfully in Cloud, including with custom
tab and route behavior.