Fix asset checks + key prefixes (#16327)
Context:
https://elementl-workspace.slack.com/archives/C05KFG9TETB/p1693931080865119?thread_ts=1693924159.102829&cid=C05KFG9TETB

This is a breaking change that will always require `@asset_check(` to
have an `asset`. I think this isn't too bad ergonomically, though it
does mean repeating in the io manager case:

```python
@asset_check(asset=my_asset)
def my_check(my_asset):
  ...
```

At present I think this is required, because we use asset groups for
resolving keys to the fully prefixed keys, and we don't know the group
of a check until we resolve which asset it points at.