Use a different, simpler Toposort priority queue when calculating get_root_unreconciled_ancestors in the asset daemon
Summary:
It has become clear that in https://github.com/dagster-io/dagster/pull/16150 I was not correct about the behavior of ToposortedPriorityQueue. I thought it was a priority queue that took in a list of asset keys, sorted them topologically, and then returned them in that order (with a list for each level. It has become clear that it actually does something more subtle, where it computes get_required_multi_asset_keys for each passed in asset key, sorts those, and returns a list for each element of those asset keys.

What that means is that in practice, the asset keys coming out of ToposortedPriorityQueue are not always in topological order (and in at least some cases seem to be a much larger set than what was passed in), causing errors since we were depending on them doing so..

We should either revert the change in https://github.com/dagster-io/dagster/pull/16150 or land this for the release.